Improve intellisense, add translation dlls in installer
