--- OpenRPA.CodeEditor/packages.config
@@ -1,31 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="AvalonEdit" version="6.0.0" targetFramework="net462" />
-  <package id="Microsoft.CodeAnalysis.Analyzers" version="2.9.8" targetFramework="net462" developmentDependency="true" />
-  <package id="Microsoft.CodeAnalysis.AnalyzerUtilities" version="3.0.0" targetFramework="net462" />
-  <package id="Microsoft.CodeAnalysis.Common" version="3.4.0" targetFramework="net462" />
-  <package id="Microsoft.CodeAnalysis.CSharp" version="3.4.0" targetFramework="net462" />
-  <package id="Microsoft.CodeAnalysis.CSharp.Features" version="3.4.0" targetFramework="net462" />
-  <package id="Microsoft.CodeAnalysis.CSharp.Workspaces" version="3.4.0" targetFramework="net462" />
-  <package id="Microsoft.CodeAnalysis.Features" version="3.4.0" targetFramework="net462" />
-  <package id="Microsoft.CodeAnalysis.FlowAnalysis.Utilities" version="2.9.6" targetFramework="net462" />
-  <package id="Microsoft.CodeAnalysis.VisualBasic" version="3.4.0" targetFramework="net462" />
-  <package id="Microsoft.CodeAnalysis.VisualBasic.Features" version="3.4.0" targetFramework="net462" />
-  <package id="Microsoft.CodeAnalysis.VisualBasic.Workspaces" version="3.4.0" targetFramework="net462" />
-  <package id="Microsoft.CodeAnalysis.Workspaces.Common" version="3.4.0" targetFramework="net462" />
-  <package id="Microsoft.DiaSymReader" version="1.3.0" targetFramework="net462" />
-  <package id="System.Buffers" version="4.5.0" targetFramework="net462" />
-  <package id="System.Collections.Immutable" version="1.7.0" targetFramework="net462" />
-  <package id="System.Composition" version="1.4.0" targetFramework="net462" />
-  <package id="System.Composition.AttributedModel" version="1.4.0" targetFramework="net462" />
-  <package id="System.Composition.Convention" version="1.4.0" targetFramework="net462" />
-  <package id="System.Composition.Hosting" version="1.4.0" targetFramework="net462" />
-  <package id="System.Composition.Runtime" version="1.4.0" targetFramework="net462" />
-  <package id="System.Composition.TypedParts" version="1.4.0" targetFramework="net462" />
-  <package id="System.Memory" version="4.5.3" targetFramework="net462" />
-  <package id="System.Numerics.Vectors" version="4.6.0-preview5.19224.8" targetFramework="net462" />
-  <package id="System.Reflection.Metadata" version="1.8.0" targetFramework="net462" />
-  <package id="System.Runtime.CompilerServices.Unsafe" version="4.7.0" targetFramework="net462" />
-  <package id="System.Text.Encoding.CodePages" version="4.7.0" targetFramework="net462" />
-  <package id="System.Threading.Tasks.Extensions" version="4.6.0-preview.18571.3" targetFramework="net462" />
-</packages>
