--- OpenRPA.ExpressionEditor/AutoCompletionPopup.xaml
@@ -1,27 +1,0 @@
-﻿<ToolTip x:Class="OpenRPA.ExpressionEditor.AutoCompletionPopup"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:OpenRPA.ExpressionEditor"
-             mc:Ignorable="d" 
-             d:DesignHeight="450" d:DesignWidth="800">
-    <Grid Name="MainGrid" Width="300" Height="200" >
-        <ListBox Name="AutoCompletionListBox" IsTextSearchEnabled="True" 
-        ItemsSource="{Binding Path=.}" VirtualizingStackPanel.IsVirtualizing="True"
-        ScrollViewer.VerticalScrollBarVisibility="Auto"
-        ScrollViewer.HorizontalScrollBarVisibility="Disabled">
-            <ListBox.ItemContainerStyle>
-                <Style TargetType="{x:Type ListBoxItem}">
-                    <EventSetter Event="MouseDoubleClick" Handler="OnListBoxItemDoubleClick" />
-                </Style>
-            </ListBox.ItemContainerStyle>
-            <ListBox.ItemTemplate>
-                <DataTemplate>
-                    <TextBlock Width="Auto" Text="{Binding Path=Name}"
-                ToolTip="{Binding Path=Description}" Margin="2" />
-                </DataTemplate>
-            </ListBox.ItemTemplate>
-        </ListBox>
-    </Grid>
-</ToolTip>
