--- OpenRPA.ExpressionEditor/OpenRPA.ExpressionEditor.csproj
@@ -1,75 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFramework>net462</TargetFramework>
-    <Authors>OpenRPA</Authors>
-    <Product>OpenRPA</Product>
-    <PackageLicenseExpression>MPL-2.0</PackageLicenseExpression>
-    <PackageProjectUrl>https://github.com/open-rpa/openrpa</PackageProjectUrl>
-    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <Version>1.0.9</Version>
-    <PackageIcon>open_rpa128.png</PackageIcon>
-    <PackageReleaseNotes>1.0.8 update interfaces reference
-1.0.5 update interfaces reference</PackageReleaseNotes>
-    <Configurations>Debug;Release;ReleaseNuget;PrepInstaller</Configurations>
-  </PropertyGroup>
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
-    <OutputPath>..\dist</OutputPath>
-  </PropertyGroup>
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='PrepInstaller|AnyCPU'">
-    <OutputPath></OutputPath>
-  </PropertyGroup>
-  <Target Name="PushNugetPackage" AfterTargets="Pack" Condition="'$(Configuration)' == 'ReleaseNuget'">
-    <Exec Command="nuget.exe push $(OutputPath)..\$(PackageId).$(PackageVersion).nupkg -Source nuget.org" />
-  </Target>
-  <ItemGroup>
-    <None Include="Resources\open_rpa128.png" Pack="true" PackagePath="\" />
-  </ItemGroup>
-  <ItemGroup>
-    <None Remove="AutoCompletionPopup.xaml" />
-    <None Remove="Images\class.png" />
-    <None Remove="Images\delegate.png" />
-    <None Remove="Images\enum.png" />
-    <None Remove="Images\event.png" />
-    <None Remove="Images\field.png" />
-    <None Remove="Images\interface.png" />
-    <None Remove="Images\method.png" />
-    <None Remove="Images\namespace.png" />
-    <None Remove="Images\property.png" />
-    <None Remove="Images\valuetype.png" />
-  </ItemGroup>
-  <ItemGroup>
-    <Page Include="AutoCompletionPopup.xaml">
-      <Generator>MSBuild:Compile</Generator>
-    </Page>
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\OpenRPA.Interfaces\OpenRPA.Interfaces.csproj" />
-  </ItemGroup>
-  <ItemGroup>
-    <Reference Include="PresentationCore" />
-    <Reference Include="PresentationFramework" />
-    <Reference Include="System.Activities" />
-    <Reference Include="System.Activities.Presentation" />
-    <Reference Include="System.Xaml" />
-    <Reference Include="WindowsBase" />
-  </ItemGroup>
-  <ItemGroup>
-    <Resource Include="Images\class.png" />
-    <Resource Include="Images\delegate.png" />
-    <Resource Include="Images\enum.png" />
-    <Resource Include="Images\event.png" />
-    <Resource Include="Images\field.png" />
-    <Resource Include="Images\interface.png" />
-    <Resource Include="Images\method.png" />
-    <Resource Include="Images\namespace.png" />
-    <Resource Include="Images\property.png" />
-    <Resource Include="Images\valuetype.png" />
-  </ItemGroup>
-  <Target Name="PostBuild" AfterTargets="Pack" Condition="'$(Configuration)' == 'Debug'">
-    <Exec Command="copy $(OutputPath)..\$(PackageId).$(PackageVersion).nupkg C:\code\OpenRPA\packages" />
-  </Target>
-  <Target Name="PostBuild" AfterTargets="Pack" Condition="'$(Configuration)' == 'PrepInstaller'">
-    <Exec Command="copy $(OutputPath)..\$(PackageId).$(PackageVersion).nupkg C:\code\OpenRPA\packages" />
-  </Target>
-  <ProjectExtensions><VisualStudio><UserProperties BuildVersion_StartDate="2000/1/1" /></VisualStudio></ProjectExtensions>
-</Project>
