+++ OpenRPA/OpenRPA.csproj
@@ -9,8 +9,9 @@
     <Description>Base UI of OpenRPA</Description>
     <PackageLicenseExpression>MPL-2.0</PackageLicenseExpression>
     <PackageProjectUrl>https://github.com/open-rpa/openrpa</PackageProjectUrl>
-    <Version>1.0.61</Version>
-    <PackageReleaseNotes>1.0.61 Finish implementing intellisense, add preloading of intellisense
+    <Version>1.0.62</Version>
+    <PackageReleaseNotes>1.0.62 Improve intellisense, add translation dlls in installer
+1.0.61 Finish implementing intellisense, add preloading of intellisense
 1.0.60 Begin implementing intellisense for C# and VB
 1.0.59 Make enable_cache default true
 1.0.58 Improve UI and speed
